Add placeholders for board, players & controller
Created initial objects for the game board, list of players and the
controller that is responsible for controlling game flow on the
display. All of them are placeholders for the time being until more
thought is put into the functionality and how to get things to work on
console before styling it up.