Rework gameboard function
Revised array to closer reflect what I imagined the array setup to be in
the function. Set up a function to display the current state of the
"Board" (the array). Also set up functions within the factory that
allow players to take their turn and to check and see if there is a
game winner each turn.