Modify Gameboard function to fix IIFE elements
Changed all function declarations into function expressions as well as
added necessary return statement for all expressions as console was not
recognizing anything as defined prior. Relocated win/loss console
messages into checkWin function in an attempt to fix the error of the
code not recognizing a winner, but error still persists. Game is
currently playable on console but no winner can be declared. Added
condition to check for tie games but has not been tested yet. Made dev
comments reflecting current status of the game.