Modify checkWin function to fix winner errors
Modified the code within the checkWin function as the console was not
displaying who won the game even after any 3-in-a-row; horizontal,
diagonal or vertical. Removed all previous for loops and went with a
for...of loop with a switch statement as the working code for checkWin.
Lowered the scope of the "marker" variable and placed within the
checkWin function in order to get the function to work properly and
actually declare a winner. Removed dev notes related to checkWin
function that were no longer needed.