Modify and reorganize HTML and styling
Reorganized HTML as it was in the wrong spot before (the head instead of
the body). Added the 3 x 3 grid representing the UI for the game board
in CSS and gave it some styling. Restyled the "Start" and "Restart"
buttons. Still need to position both the grid and the buttons correctly.
Will add space for Player 1 and Player 2 names above the game board.