Add disable functionality to displayController
Implemented a new disableAll function within displayController that will
disable all buttons and cells on the board once a winner/loser is
declared. Added the function as comments within Gameboard for the time
being in the necessary places (checkWin function) to test out later on.
Added more general dev comments on possible next steps and code changes
throughout the JS.