Modify Gameboard and displayController functions
Modified the makeMove function inside the Gameboard by removing the
plural that represented all indexes in the markers array and replacing
it with singular 'marker' so we can use a specific marker (the current
player's turn) as an argument for the cell click events in the
displayController function. Revised the displayAll function in the
displayController to include a conditional 'if' statement since simply
setting attributes to disabled directly may not work once everything is
debugged. Also made the exact same change to the 'cells' click event
listener. Re-enabled the displayAll code in the checkWin function inside
of Gameboard (currently). Added more dev comments regarding possible
next steps - either gameController becoming an IIFE or reorganzing all
functionality between it and the Gameboard (and possibly the
displayController as well).