Modify marker code for debugging
Attempted to change all instances of 'X' and 'O' strings as markers to
target the array inside the checkWin function instead in order to
resolve the markers showing as undefined at game start and when clicking
a cell on the board. Attempt was unsuccessful. Reverted back to previous
code and added more dev comments regarding attempt and some next steps.