Modify Players function in a debug attempt
Attempted several different solutions to try and get player markers to
both assign correctly and show as 'X' and 'O' instead of 'undefined'.
Rewrote the Players function as a factory function and moved it to the
global scope while moving the random marker assignment from startGame
to the new factory function to no avail. Reverted all Players function
code back to it's previous state. Made dev comments regarding the
attempts as well as notes on why switchTurn continuously runs in the
switch statement under the checkWin function after a cell is clicked,
among other notes.