Remove randomized marker assignment option
Removed all code that tried to assign the first marker to the first
player randomly at the start of the game after several failed debugging
attempts. Now 'X' always goes first. Replaced any functions named
'Player' or 'createPlayer' with a simple object containing both players
as properties with values. Refactored other code due to initialization
errors I was getting after testing. Removed all dev comments relating
to markers showing as undefined as that had to do with the randomized
markers not being recognized. May try to add this feature to the new
'players' object in the future, but no longer a priority.