Add comments & modify DOM in gameController
Modified the player1 and player2 DOM elements in an attempt to get the
console to display the names that the players entered correctly to no
avail. Added comments throughout the code as a reminder that this still
needs to be fixed as well as deleting certain console.logs once the UI
is working correctly during certain parts of the game.