Modify comments related to player name bug
Updated & deleted some comments related to the player's entered name not
showing up in the console when the turn is switched after a cell is
chosen. The bug seems to be either in the DOM element itself or the
object containing the player's info (their name and marker). In either
case, the value that was entered by the player in the name box area is
not being read for some reason (the app reads an empty string instead).
Tested to make sure both markers in the object are being seen correctly
by playing 2 turns of the game and can confirm they are working as
intended.