Fix empty string error when switching turns
Modified both the player1 and player2 DOM variables as well as the
setNames methods inside the players array in gameController. The methods
now return the value of what was entered instead of the value being tied
to the DOM variable directly. The name the user entered now prints
correctly in the console when switching turns. Will experiment with
turning the name properties into the exact same method as setNames as an
attempt to refactor and clean up the code a bit. If it works, original
setNames methods would be deleted, reducing the amount of keys in each
object under players array.