Add all UI console messages as dynamic info text
Added functionality in displayController for the UI to show information
on whose turn it is and which cells become occupied during a given turn.
Info shows in a header element. Player name is not showing at all -
similar error to when the name wasn't showing on console. However, the
name IS still showing correctly in the console, so it is a separate
issue in the code itself (not how I'm calling the name input). Player
array in gameController has also been refactored. Comments that had
completed attempts to fix bugs have been removed.