Comment possible UI winner display in checkWin
Added a possible solution to displaying the winner of the game in the
UI under the checkWin function (inside of column check). Modified other
comments related to this function reflecting this. Added "info" to list
of return values under displayController so checkWin can access it to
show the winner messages.