Modify checkWin function & UI display comments
Updated the "tie game check" for the switch statement in the checkWin
function. Tie game scenario now runs correctly but there is still an
issue with the UI showing the next players turn even after the game
ends. Updated comments in displayController reflecting possible next
steps in the forEach method for the board cells.