Modify switch turn logic for the cells in the UI
Placed all the switch turn logic and display messages in the cell event
listeners under a conditional statement that checks first to see if
there is a game winner from checkWin. This should resolve the bug where
the player's turn would show on the info header in the UI when the game
ended instead of who won the game. There is still an issue of the
winning/draw message not showing when the game ends in checkWin. Added
comments describing what the "position" messages does and shows under
what conditions in displayController.