Resolve switch turn message error at game end
Modified second conditional statement in cells functionality as it was
not correctly checking the state of checkWin in the Gameboard function.
Code now correctly switches turns and displays the correct players turn
in the UI as long as there is not a game winner. If there is, code
correctly omits the player turn message and the turn switching so the
last game state and UI message will be the final game result. Removed
comments related to TODO and OPTION that have either been completed or
are no longer needed.