Modify then revert Restart game functionality
Attempted to modify all restart game functionality with location.reload
as a one size fits all solution. Unfortunately, it causes too much
flickering on the webpage to the point where users can't even interact
with it. The page seems to continuously refresh without stopping. Will
need to attempt to find another solution to the Restart game/button
functionality not working as intended.