Resolved location.reload flickering
The parentheses inside of the restartBtn when it called restartGame was
causing all the flickering and making the page non-interactive. Removed
the parentheses on the button and the window.stop method on the
restartGame function as it was no longer needed. Restart button now
functions correctly (fully refreshes the page and resets everything
with no flickering). Deleted all comments in HTML, CSS and JS files
that were no longer needed.